qat17: add support for DISTRO_FEATURE usrmerge
Issue: LINUXEXEC-31579

The build script of qat17_4.24.0-00005 will put the firmware and
drivers under <path of rootfs>/lib. This is hardcoded in QAT's makefile.

When "usrmerge" is in DISTRO_FEATURES, "nonarch_base_libdir" is changed
to "/usr/lib", the recipe can install firmware to correct location.
But we need to move all qat drivers to "/usr/lib/modules/<kernel version>".

(LOCAL REV: NOT UPSTREAM) - will sent to meta-intel-qat later

Signed-off-by: Yongxin Liu <yongxin.liu@windriver.com>